+++ src/main/java/org/sourcelab/kafka/connect/apiclient/request/post/PostConnectorRestart.java
@@ -17,16 +17,22 @@
 
 package org.sourcelab.kafka.connect.apiclient.request.post;
 
-import java.io.IOException;
-import java.util.Objects;
+import org.sourcelab.kafka.connect.apiclient.util.UrlEscapingUtil;
 
-import static com.google.common.net.UrlEscapers.urlPathSegmentEscaper;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 
 /**
  * Defines a request to restart a connector.
  */
 public final class PostConnectorRestart implements PostRequest<Boolean> {
     private final String connectorName;
+
+    private Map<String, Boolean> options = new HashMap<>();
 
     /**
      * Constructor.
@@ -37,9 +43,48 @@ public final class PostConnectorRestart implements PostRequest<Boolean> {
         this.connectorName = connectorName;
     }
 
+    /**
+     * Only available from Kafka Connect version 3.0.0 and up.
+     * @param includeTasks Specifies whether to restart the connector instance and task instances (includeTasks=true`) or
+     *                     just the connector instance (includeTasks=false).
+     * @return self reference for method chaining.
+     */
+    public PostConnectorRestart withIncludeTasks(final boolean includeTasks)
+    {
+        options.put("includeTasks", includeTasks);
+        return this;
+    }
+
+    /**
+     * Only available from Kafka Connect version 3.0.0 and up.
+     * @param onlyFailed specifies whether to restart just the instances with a FAILED status (onlyFailed=true)
+     *                   or all instances (onlyFailed=false).
+     * @return self reference for method chaining.
+     */
+    public PostConnectorRestart withOnlyFailed(final boolean onlyFailed)
+    {
+        options.put("onlyFailed", onlyFailed);
+        return this;
+    }
+
     @Override
     public String getApiEndpoint() {
-        return "/connectors/" + urlPathSegmentEscaper().escape(connectorName) + "/restart";
+        // Define base URL
+        String url = "/connectors/" + UrlEscapingUtil.escapePath(connectorName) + "/restart";
+
+        // Optionally add additional request parameters if explicitly defined.
+        final List<String> params = new ArrayList<>();
+        for (final Map.Entry<String, Boolean> option : options.entrySet()) {
+            // skip null
+            if (option.getValue() == null) {
+                continue;
+            }
+            params.add(option.getKey() + "=" + (option.getValue() ? "true" : "false"));
+        }
+        if (params.size() > 0) {
+            url = url + "?" + String.join("&", params);
+        }
+        return url;
     }
 
     @Override
@@ -49,6 +94,7 @@ public final class PostConnectorRestart implements PostRequest<Boolean> {
 
     @Override
     public Boolean parseResponse(final String responseStr) throws IOException {
+        // Note: this doesn't currently support 202 responses which would normally contain a response body :/
         return true;
     }
 }