+++ src/main/java/org/sourcelab/kafka/connect/apiclient/util/UrlEscapingUtil.java
@@ -15,41 +15,30 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package org.sourcelab.kafka.connect.apiclient.request;
+package org.sourcelab.kafka.connect.apiclient.util;
 
-import org.apache.commons.io.IOUtils;
-import org.junit.Test;
-
-import java.io.IOException;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
+import org.apache.http.client.utils.URLEncodedUtils;
+import java.util.Objects;
 
 /**
- * Abstract / Base Request Test Class.
+ * Intended to provide a shim between Guava so that dependency can be removed in a future release.
  */
-abstract public class AbstractRequestTest {
-
+public class UrlEscapingUtil {
     /**
-     * Test Parsing response.
+     * Returns the escaped form of a given literal string.
+     *
+     * @param input the literal string to be escaped.
+     * @return the escaped form of string.
+     * @throws NullPointerException if input string is null.
+     * @throws IllegalArgumentException if string contains badly formed UTF-16 or cannot be escaped for any other reason
      */
-    @Test
-    public abstract void testParseResponse() throws Exception;
-
-    /**
-     * Verifies the behavior of the getApiEndPoint to properly return an escaped URL
-     * when the connector name requires escaping.
-     */
-    @Test
-    public abstract void getApiEndpoint();
-
-    /**
-     * Utility method to help load mock responses from resources.
-     * @param fileName file name to load from resources
-     * @return The contents of the file, as a UTF-8 string.
-     * @throws IOException on error reading from resource file.
-     */
-    protected String readFile(final String fileName) throws IOException {
-        final URL inputFile = getClass().getClassLoader().getResource("mockResponses/" + fileName);
-        return IOUtils.toString(inputFile, StandardCharsets.UTF_8);
+    public static String escapePath(final String input) {
+        Objects.requireNonNull(input);
+        final String result = URLEncodedUtils.formatSegments(input);
+        // Strip prepended slash
+        if (result.length() > 0) {
+            return result.substring(1);
+        }
+        return "";
     }
 }