+++ src/test/java/org/sourcelab/kafka/connect/apiclient/request/get/connector/GetConnectorsExpandAllDetailsTest.java
@@ -188,4 +188,11 @@ public class GetConnectorsExpandAllDetailsTest extends AbstractRequestTest {
         assertEquals("trace0", connectorStatus.getTasks().get(0).getTrace());
         assertEquals("127.0.0.1:8083", connectorStatus.getTasks().get(0).getWorkerId());
     }
+
+    @Override
+    public void getApiEndpoint() {
+        final String expectedUrl = "/connectors?expand=info&expand=status";
+        final String result = new GetConnectorsExpandAllDetails().getApiEndpoint();
+        assertEquals("Unexpected URL returned!", expectedUrl, result);
+    }
 }